@@ -1,19 +1,17 @@
-'use strict';
+const rest = require('rest');
+const defaultRequest = require('rest/interceptor/defaultRequest');
+const mime = require('rest/interceptor/mime');
+const uriTemplateInterceptor = require('./api/uriTemplateInterceptor');
+const errorCode = require('rest/interceptor/errorCode');
+const baseRegistry = require('rest/mime/registry');
 
-var rest = require('rest');
-var defaultRequest = require('rest/interceptor/defaultRequest');
-var mime = require('rest/interceptor/mime');
-var uriTemplateInterceptor = require('./api/uriTemplateInterceptor');
-var errorCode = require('rest/interceptor/errorCode');
-var baseRegistry = require('rest/mime/registry');
-
-var registry = baseRegistry.child();
+const registry = baseRegistry.child();
 
 registry.register('text/uri-list', require('./api/uriListConverter'));
 registry.register('application/hal+json', require('rest/mime/type/application/hal'));
 
 module.exports = rest
-		.wrap(mime, { registry: registry })
-		.wrap(uriTemplateInterceptor)
-		.wrap(errorCode)
-		.wrap(defaultRequest, { headers: { 'Accept': 'application/hal+json' }});
+  .wrap(mime, {registry})
+  .wrap(uriTemplateInterceptor)
+  .wrap(errorCode)
+  .wrap(defaultRequest, {headers: {'Accept': 'application/hal+json'}});
